batch: Move line sequence search
match.py computes conservative source-to-target line mappings for batch
operations.

Exact sequence occurrence scans and context-gap searches are a separate
placement concern used by presence placement. Keeping them in the matcher
made its public surface mix mapping with local search.

This commit addresses that by moving TargetGap and exact line sequence
search helpers into line_sequence_search and routing presence placement
through it.

Subsequent architecture coverage will record the matcher boundary.
